Fix `#finish` when connected but not yet started
Because `#started?` returns false until _after_ authentication,
authenticators can't safely call `finish`.  The `IOError` was raised
before the socket could be closed.

An authenticator may need to close the connection:
* if it supports mutual authentication and that has failed.
* if the server is behaving in an untrustworthy way, for example
  claiming that authentication has succeeded before the authentication
  exchange has completed).

Because these represent failures from the server, we should not be
concerned with sending `QUIT`, even if it might be possible.